refactor: change folder name to explore
