+++ front-end/src/routes.tsx
@@ -4,7 +4,7 @@ import Dashboard from "./pages/private/Dashboard";
 import Login from "./pages/public/Login";
 import ProtectedLayout from "./components/ProtectedLayout";
 import Onboarding from "./pages/private/Onboarding";
-import Explore from "./pages/private/Search";
+import Explore from "./pages/private/Explore";
 
 export const router = createBrowserRouter([
   {
